% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dys_pvcm.R
\name{dys_pvcm}
\alias{dys_pvcm}
\title{Davis-Yin Splitting for PVCM}
\usage{
dys_pvcm(
  y_vec,
  M_mat,
  tau,
  Theta_initial,
  weights_ann,
  weights_agl,
  gamma_val,
  max.iter = 100,
  tol = 1e-05
)
}
\arguments{
\item{y_vec}{A numeric vector of response variable}

\item{M_mat}{A matrix whose row is the kroneker product between the covariates X_i and the evaluated basis functions at U_i}

\item{tau}{A vector of the two tuning parameters}

\item{Theta_initial}{An initial value for the Theta matrix}

\item{weights_ann}{A vector of weights for the adaptive nuclear norm penalty}

\item{weights_agl}{A vector of weights for the adaptive group lasso penalty}

\item{gamma_val}{A numeric value for one over Lipschitz constant}

\item{max.iter}{A maximum iteration number}

\item{tol}{A numeric value for tolerance of relative deviation in order to stop DYS iteration}
}
\value{
A list that contains estimated Theta matrices with different characteristics
}
\description{
Davis-Yin Splitting for PVCM
}
