% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pvcm.R
\name{plot_pvcm}
\alias{plot_pvcm}
\title{Trajectory plots for vayring coefficient functions and principal functions}
\usage{
plot_pvcm(pvcm_fit, pallet_size = c(1, 1), U_sample, beta_plot = TRUE)
}
\arguments{
\item{pvcm_fit}{A list output of pvcm function}

\item{pallet_size}{A vector of two positive integer elements for the number of rows and columns of pallet}

\item{U_sample}{A numeric vector of index variable}

\item{beta_plot}{A logical value to indicate whether the coefficient functions (TRUE) or the principal functions (FALSE) are displayed}
}
\description{
Trajectory plots for vayring coefficient functions and principal functions
}
