% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvcm.R
\name{pvcm}
\alias{pvcm}
\title{Principal Varying Coefficient Model with Adaptive weights}
\usage{
pvcm(
  y_sample,
  X_sample,
  U_sample,
  num_cores = NULL,
  knots_val = 7,
  control = list(num_tau_1 = 20, num_tau_2 = 20, max.iter = 100, tol = 1e-05)
)
}
\arguments{
\item{y_sample}{A numeric vector of response variable}

\item{X_sample}{A design matrix of covariates including an intercept if it is included in your model}

\item{U_sample}{A numeric vector of index variable}

\item{num_cores}{The number of CPU cores to implement parallel computing (Default NULL will use the total number of cores - 1)}

\item{knots_val}{A numeric value for the knots argument of select.knots function in the face R package}

\item{control}{A list that contains two numbers of tuning parameters except 0 and the stopping criteria for DYS algorithm}
}
\value{
A list that contains all estimation results
}
\description{
Principal Varying Coefficient Model with Adaptive weights
}
